refactor(tui): split scrollback logic out of terminal-compositor.types.ts (#826)
terminal-compositor.types.ts was named `.types.ts` but held 7 function
declarations (~310 LOC) carrying real scrollback-flush and ANSI-escape
logic, most with dense #540-incident Invariant/History commentary. A
reviewer scanning for "the file with the scrollback algorithm" would not
look in a file promising zero logic.

Move all 7 functions verbatim to a new terminal-compositor.scrollback.ts:
formatElapsed, eraseAndPaintRow, buildBandMeta, scrollbackFlushLines,
snapFlushCountToLogicalBoundary, buildScrollbackArchiveEscape,
formatTipRow (plus their sole constant dependency, ELAPSED_GRACE_MS).
terminal-compositor.types.ts keeps only interfaces/types/consts, finally
honest about its name. BandRowMeta stays in .types.ts (not moved)
because terminal-compositor.ts imports it directly and that file is out
of scope for this change; terminal-compositor.scrollback.ts imports it
back as a type-only import.

All 7 functions were confirmed pure by grep (no `self: Host` parameter,
unlike every stateful sibling module in the compositor cluster) before
the move, matching the issue's claim.

Import sites updated directly (no re-export shim) at the 8 places that
imported one of the 7 functions from the old path: input/spinner.ts,
terminal-compositor.committed-band-{repin,commit}.ts,
terminal-compositor.frame-preserve-archive.ts,
terminal-compositor.lifecycle.ts, and 3 test files. A re-export would
have left the exact `.types.ts` import path issue #826 complains about
fully reachable, defeating the discoverability fix; the compiler
verifies every updated site (`tsc --noEmit` is loud on a missed one).

Corrected premise: the issue claims all 7 functions have "zero tests".
False as of PR #665 (2026-07-28, pre-dates this issue) — 6 of 7 already
have direct coverage:
  buildBandMeta, scrollbackFlushLines, snapFlushCountToLogicalBoundary,
    buildScrollbackArchiveEscape -> terminal-compositor.logical-flush.test.ts
  formatTipRow                   -> loading-tips.test.ts
  eraseAndPaintRow                -> terminal-compositor.render-not-repin.test.ts,
                                     terminal-compositor.resize-stale-width.repro.test.ts
formatElapsed was the sole real gap; added
terminal-compositor.format-elapsed.test.ts (8 cases: grace-period gate,
seconds-only vs. minutes+seconds format switch at the 60s boundary, and
zero-padding).

No behavior change — bodies are byte-for-byte moves (diffed against
git-show HEAD to confirm).

Closes #826
